Commit - Spring Boot project setup 8 (JWT Authentication) (Register user, Password encrypted,Login dari database, Generate JWT, JWT Authentication, Protected API)

+++ src/main/java/com/example/employee_api/repository/UserRepository.java
@@ -2,9 +2,10 @@ package com.example.employee_api.repository;
 
 import com.example.employee_api.entity.User;
 import org.springframework.data.jpa.repository.JpaRepository;
+
 import java.util.Optional;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
     Optional<User> findByUsername(String username);
 
@@ -15,8 +16,10 @@ public interface UserRepository extends JpaRepository<User, Integer> {
 //
 //import com.example.employee_api.entity.User;
 //import org.springframework.data.jpa.repository.JpaRepository;
-//import java.util.*;
+//import java.util.Optional;
 //
-//public interface UserRepository extends JpaRepository<User, Long> {
+//public interface UserRepository extends JpaRepository<User, Integer> {
+//
 //    Optional<User> findByUsername(String username);
+//
 //}
